+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/mapper/MinMapperSupporter.java
@@ -31,6 +31,6 @@ public class MinMapperSupporter implements MapperSupporter {
 
     @Override
     public String provideArgsIllustrate() {
-        return "";
+        return "不需要任何参数";
     }
 }